+++ include/gen_key.h
@@ -1,8 +1,8 @@
 #ifndef GEN_KEY_H
 #define GEN_KEY_H
 
-void export_rsa_pub_key(int e, int n);
+void export_rsa_pub_key(long long e, long long n);
 
-void export_rsa_priv_key(int d, int n);
+void export_rsa_priv_key(long long d, long long n);
 
 #endif // GEN_KEY_H
