+++ include/rsa_gen.h
@@ -1,8 +1,8 @@
 #ifndef RSA_GEN_H
 #define RSA_GEN_H
 
-int rsa_pub(int m, int e, int n);
+long long rsa_pub(long long m, long long e, long long n);
 
-int rsa_priv(int m, int d, int n);
+long long rsa_priv(long long m, long long d, long long n);
 
 #endif // RSA_GEN_H
