+++ src/rsa_gen.c
@@ -1,11 +1,11 @@
 #include "../include/utils.h"
 
-int rsa_pub(int m, int e, int n)
+long long rsa_pub(long long m, long long e, long long n)
 {
     return mod_exp(m, e, n);
 }
 
-int rsa_priv(int m, int d, int n)
+long long rsa_priv(long long m, long long d, long long n)
 {
     return mod_exp(m, d, n);
 }
